GEODE-3473: Initial commit of the internal package renaming refactor. Changing the protobuf file refactor first.
