+++ geode-protobuf/src/main/java/org/apache/geode/protocol/protobuf/Result.java
@@ -17,6 +17,7 @@ package org.apache.geode.protocol.protobuf;
 import java.util.function.Function;
 
 import org.apache.geode.annotations.Experimental;
+import org.apache.geode.internal.protocol.protobuf.BasicTypes;
 
 @Experimental
 public interface Result<SuccessType> {
