+++ geode-protobuf/src/main/proto/authentication_API.proto
@@ -14,7 +14,7 @@
  */
 
 syntax = "proto3";
-package org.apache.geode.protocol.protobuf;
+package org.apache.geode.internal.protocol.protobuf;
 
 message SimpleAuthenticationRequest {
     string username = 1;
