+++ geode-protobuf/src/main/proto/clientProtocol.proto
@@ -19,7 +19,7 @@
 */
 
 syntax = "proto3";
-package org.apache.geode.protocol.protobuf;
+package org.apache.geode.internal.protocol.protobuf;
 
 import "google/protobuf/any.proto";
 import "region_API.proto";
