+++ geode-protobuf/src/main/proto/region_API.proto
@@ -19,7 +19,7 @@
 */
 
 syntax = "proto3";
-package org.apache.geode.protocol.protobuf;
+package org.apache.geode.internal.protocol.protobuf;
 
 import "basicTypes.proto";
